requirements.txt
langchain-groq
fastapi
uvicorn
python-dotenv
pandas
tqdm
requests
bs4
lxml
chardet
streamlit
PyPDF2
python-docx
pyperclip